requirements.txt
# requirements.txt content
REQUIREMENTS = """
MetaTrader5==5.0.45
pandas==2.0.3
numpy==1.24.3
schedule==1.2.0
python-dotenv==1.0.0
"""